% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFsamples.R
\name{plotFsamples}
\alias{plotFsamples}
\title{Plot one-dimensional probability distributions}
\usage{
plotFsamples(
  file,
  mcoutput,
  data,
  plotmeans = TRUE,
  plotuncertainty = "samples",
  uncertainty = 100,
  datahistogram = TRUE,
  datascatter = TRUE,
  useOquantiles = FALSE,
  parallel = TRUE,
  silent = FALSE
)
}
\arguments{
\item{mcoutput}{Either a string with the name of a directory or full path
for a 'FDistribution.rds' object, or such an object itself}

\item{silent}{Bool: give warnings or updates in the computation}
}
\description{
Plot one-dimensional probability distributions
}
